+++ feature/explore/src/main/java/com/teamoffroad/feature/explore/domain/repository/QuestRepository.kt
@@ -3,6 +3,9 @@ package com.teamoffroad.feature.explore.domain.repository
 import com.teamoffroad.feature.explore.domain.model.Quest
 
 interface QuestRepository {
-
-    suspend fun fetchQuests(isActive: Boolean, cursor: Int, size: Int): List<Quest>
+    suspend fun fetchQuests(
+        isActive: Boolean,
+        cursor: Long,
+        size: Int,
+    ): List<Quest>
 }
