+++ feature/explore/src/main/java/com/teamoffroad/feature/explore/presentation/model/QuestModel.kt
@@ -5,7 +5,7 @@ data class QuestModel(
     val description: String,
     val requirement: String,
     val reward: String,
-    val cursorId: Int,
+    val cursorId: Long,
     val questProgressModel: QuestProgressModel,
 ) {
     data class QuestProgressModel(
